adding read and write to can using struct

+++ src/displayHandler/display_class.cpp
@@ -10,7 +10,7 @@ void displayHandler::initDisplay()
 
 void displayHandler::update()
 {
-    std::cout << std::dec << std::setw(3) << std::setfill('0') << "Accelerator: " << (uint8_t)accelerator << " %" << std::endl;
-    std::cout << std::dec << std::setw(4) << std::setfill('0') << "Brake:       " << (uint8_t)brake << " %" << std::endl;
-    std::cout << std::dec << std::setw(4) << std::setfill('0') << "EngineSpeed: " << (uint16_t)rpm << " rpm" << std::endl;    
+    std::cout << std::dec << std::setw(3) << std::setfill('0') << "Accelerator: " << (int)accelerator << " %" << std::endl;
+    std::cout << std::dec << std::setw(4) << std::setfill('0') << "Brake:       " << (int)brake << " %" << std::endl;
+    std::cout << std::dec << std::setw(4) << std::setfill('0') << "EngineSpeed: " << (int)rpm << " rpm" << std::endl;    
 }
